refactor(#27): 디렉토리 구조 변경 - 불필요한 파라미터 제거 - RedisDto 매핑 확장
