+++ README.md
@@ -1,2 +1,28 @@
 # Monomat-BE
+
 🕹️ Monomat 백엔드 리포지토리입니다.
+
+다음은 현재 도메인 중심 패키지 구조입니다.
+
+```
+src/main/java/io/github/ascrew/monomatbe/
+  chat/
+    controller/
+  lobby/
+    controller/
+    domain/
+    dto/
+    port/
+    repository/
+    service/
+    websocket/
+  messaging/
+    dto/
+    redis/
+  common/
+    annotation/
+    config/
+      security/
+    constant/
+    websocket/
+```
