+++ src/main/java/io/github/ascrew/monomatbe/common/annotation/SessionUuid.java
@@ -1,0 +1,12 @@
+package io.github.ascrew.monomatbe.common.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface SessionUuid {
+}
+
