+++ src/main/java/io/github/ascrew/monomatbe/common/config/security/SecurityConfigProd.java
@@ -1,0 +1,33 @@
+package io.github.ascrew.monomatbe.common.config.security;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+
+
+@Configuration
+@EnableWebSecurity
+@Profile("prod")
+public class SecurityConfigProd {
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http
+            .csrf(AbstractHttpConfigurer::disable)      // CSRF 보호 비활성화
+            .formLogin(AbstractHttpConfigurer::disable) // 폼 로그인 비활성화
+            .httpBasic(AbstractHttpConfigurer::disable) // HTTP Basic 인증 비활성화
+            .authorizeHttpRequests(auth -> auth
+                    // 향후 회원가입, 로그인 등 인증없이 들어와야 하는 URL이 생기면 추가
+                    // .requestMatchers("/api/auth/**").permitAll()
+                    .anyRequest().authenticated()
+
+            ); // 모든 요청에 대해 인증 필요 (운영 환경에서는 인증된 사용자만 접근 허용)
+
+        return http.build();
+    }
+}
+
