+++ src/main/java/io/github/ascrew/monomatbe/common/packageInfo.java
@@ -1,0 +1,8 @@
+package io.github.ascrew.monomatbe.common;
+
+/*
+  애플리케이션 전역에서 공통으로 사용하는 인프라/유틸리티 패키지입니다.
+
+  설정(config), 상수(constant), 어노테이션(annotation), 웹소켓 공통 처리(websocket) 등을
+  한곳에서 관리해 도메인 로직과 분리합니다.
+ */
