+++ src/main/java/io/github/ascrew/monomatbe/domain/packageInfo.java
@@ -1,9 +1,8 @@
 package io.github.ascrew.monomatbe.domain;
 
 /*
-  멀티플레이어 게임 진행을 위한 '로비(Lobby)' 데이터를 담당하는 JPA 엔티티(Entity) 클래스입니다.
+  도메인 모델의 공통 개념을 담는 패키지입니다.
 
-  데이터베이스 테이블과 1:1로 매핑되는 핵심 비즈니스 모델입니다.
-  단순한 데이터 구조체를 넘어, 입장 인원 제한, 게임 시작 상태 변경 등
-  객체 지향적으로 상태와 행위를 캡슐화하여 도메인의 핵심 비즈니스 규칙을 수호합니다.
+  각 도메인별 하위 패키지(lobby 등)에서 비즈니스 규칙과 상태 전이를 정의하며,
+  인프라 계층과 분리된 순수 도메인 타입을 유지합니다.
  */
