--- src/main/java/io/github/ascrew/monomatbe/global/annotation/SessionUuid.java
@@ -1,11 +1,0 @@
-package io.github.ascrew.monomatbe.global.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface SessionUuid {
-}
