--- src/main/java/io/github/ascrew/monomatbe/global/config/WebSocketConfig.java
@@ -1,78 +1,0 @@
-/*
-1.기본 연결 정보
-    엔드포인트(접속 주소): http://{서버_도메인_또는_IP}:8080/ws
-2.통신 경로 규칙
-    STOMP 통신 시 사용하는 경로 규칙
-    - 송신용 경로: /app/** (클라이언트가 메시지를 보낼 때 사용하는 경로)
-    - 수신용 경로: /topic/** (서버가 클라이언트에게 메시지를 보낼 때 사용하는 경로)
-3.메시지 브로커
-    - Simple Broker: /topic/** (서버가 클라이언트에게 메시지를 보낼 때 사용하는 경로)
-    - Application Destination Prefix: /app/** (클라이언트가 메시지를 보낼 때 사용하는 경로)
-4.웹소켓 연결 실패 시 대체 통신 방법
-    - SockJS: 웹소켓 연결이 실패할 경우, 폴링 등의 대체 통신 방법을 사용하여 연결을 유지할 수 있도록 지원
- */
-
-package io.github.ascrew.monomatbe.global.config;
-
-import java.util.List;
-import org.springframework.messaging.handler.invocation.HandlerMethodArgumentResolver;
-import io.github.ascrew.monomatbe.global.websocket.CustomStompErrorHandler;
-import io.github.ascrew.monomatbe.global.websocket.StompChannelInterceptor;
-import io.github.ascrew.monomatbe.global.websocket.resolver.SessionUuidArgumentResolver;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
-import org.springframework.messaging.simp.config.ChannelRegistration;
-import io.github.ascrew.monomatbe.global.constant.WebSocketConstants;
-
-
-@Configuration
-@EnableWebSocketMessageBroker
-@RequiredArgsConstructor
-public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
-
-    private final CustomStompErrorHandler customStompErrorHandler;
-    private final StompChannelInterceptor stompChannelInterceptor;
-    private final SessionUuidArgumentResolver sessionUuidArgumentResolver;
-
-    @Override
-    public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry
-                .addEndpoint(WebSocketConstants.WS_ENDPOINT) //최초로 웹소켓 연결을 하기위해 url /ws 지정
-                .setAllowedOriginPatterns("*") //모든 도메인에서 접속을 허용
-                .withSockJS(); //웹소켓 연결 실패시 일반 HTTP통신으로 연결
-
-        registry.setErrorHandler(customStompErrorHandler);
-
-    }
-
-    @Override
-    public void configureMessageBroker(MessageBrokerRegistry registry) {
-        // 하트비트를 담당할 스레드 생성 및 초기화
-        ThreadPoolTaskScheduler taskScheduler = new ThreadPoolTaskScheduler();
-        taskScheduler.setPoolSize(1);                                           //스레드 1개로 제한
-        taskScheduler.setThreadNamePrefix("wss-heartbeat-thread-");             //스레드 이름 접두사 설정
-        taskScheduler.initialize();
-
-        registry.enableSimpleBroker(WebSocketConstants.TOPIC_DESTINATION_PREFIX)                   //수신용
-                .setTaskScheduler(taskScheduler)
-                .setHeartbeatValue(new long[]{10000,10000}); //클라이언트와 서버가 10000ms(10초)마다 하트비트를 주고받도록 설정
-
-        registry.setApplicationDestinationPrefixes(WebSocketConstants.APP_DESTINATION_PREFIX);  //송신용(발행)
-    }
-
-    @Override
-    public void configureClientInboundChannel(ChannelRegistration registration) {
-        registration.interceptors(stompChannelInterceptor);
-    }
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
-        argumentResolvers.add(sessionUuidArgumentResolver);
-    }
-
-}
